fix for no uncached entry bug

+++ ece506-p3/dsm/p3/cache.h
@@ -40,7 +40,7 @@ enum dir_state {
 class dirEntry {
   public:
     ulong tag;
-    uchar state;
+    int state;
     bool bit[4];
 
     dirEntry() {
